--- domains/vms/provisioning/__init__.py
@@ -1,19 +1,0 @@
-"""VM provisioning client helpers.
-
-The executable provisioning service lives under ``domains/vms/provisioning/service``.
-Storefront-owned fulfillment, job-spec, and capacity validation helpers were
-moved into ``market_storefront`` services/models. Shared VM provision terms live
-in the ``arkhai-vms-common`` wheel.
-"""
-
-from domains.vms.provisioning.client import (
-    provision_vm_and_wait,
-    register_vm_lease,
-    schedule_vm_expiry_and_wait,
-)
-
-__all__ = [
-    "provision_vm_and_wait",
-    "register_vm_lease",
-    "schedule_vm_expiry_and_wait",
-]
