+++ domains/vms/provisioning/client/pyproject.toml
@@ -1,0 +1,25 @@
+[project]
+name = "arkhai-vms-provisioning-client"
+version = "0.1.0"
+license = "MIT"
+description = "Async and sync HTTP clients for the Arkhai provisioning service REST API."
+requires-python = ">=3.12"
+dependencies = [
+    "httpx>=0.28.1",
+    "pydantic>=2.7",
+]
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["src/provisioning_client"]
+
+[tool.pytest.ini_options]
+pythonpath = ["src"]
+
+[dependency-groups]
+dev = [
+    "pytest>=8.0",
+]
