+++ domains/vms/provisioning/service/src/controllers/hosts_controller.py
@@ -22,15 +22,15 @@ from fastapi import APIRouter, Depends, File, Form, HTTPException, UploadFile, s
 from fastapi_utils.cbv import cbv
 
 import container as _container_module
-from models.host_model import (
+from provisioning_client.models import (
     HostCreate,
     HostListResponse,
     HostResponse,
     HostUpdate,
+    JobSubmitResponse,
+    VmActionRequest,
 )
 from models.ansible import ConnectivityResult
-from models.jobs_model import JobSubmitResponse
-from models.vm_request_model import VmActionRequest
 from services.host_operations_service import HostOperationsService
 from services.host_service import HostNotFoundError, HostService
