+++ domains/vms/provisioning/service/src/services/host_operations_service.py
@@ -10,9 +10,9 @@ from __future__ import annotations
 from collections.abc import Callable
 from typing import TYPE_CHECKING
 
+from provisioning_client.models import JobSubmitResponse, VmActionRequest
 from models.ansible import ConnectivityResult
-from models.jobs_model import JobSubmitResponse
-from models.vm_request_model import VmActionRequest, build_simple_params
+from models.vm_request_model import build_simple_params
 from services.async_job_queue import AsyncJobQueue
 from services.host_service import HostNotFoundError
