+++ domains/vms/provisioning/service/src/tests/integration/test_test_controller.py
@@ -26,7 +26,7 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
 import container as _container_module
-from client.provisioning_client import ProvisioningClient, ProvisioningError
+from provisioning_client import ProvisioningClient, ProvisioningError
 from db.database import create_session_factory
 from db.models import Base
 from main import app
