+++ domains/vms/storefront/pyproject.toml
@@ -18,6 +18,7 @@ dependencies = [
     "arkhai-kit-policy",
     "arkhai-vms-common",
     "arkhai-vms-provisioning>=0.5.0",
+    "arkhai-vms-provisioning-client>=0.1.0",
     "opentelemetry-exporter-gcp-trace>=1.9.0,<2.0.0",
     "google-cloud-logging>=3.12.0,<4.0.0",
     "fastapi~=0.115.8",
